requirements.txt
torch
transformers
bitsandbytes
accelerate
datasets
pandas
numpy
matplotlib
pyyaml
tqdm
pytest
